Completes first algorithm: cardinally-adjacent neighbors.
The selection of the tile to check to see if it can grow grass is randomized. Therefore, there may be
a possible optimization in which one or more grass region(s) is(are) tracked for checking to see if a
dirt tile exists on the boundary, and then picking (at random) such a dirt tile at the boundary to
upgrade to grass. Maybe algorithm #2?