Adds the Algorithm class.
Basically, starting to extract the visualization and map data out of
the algorithm employed. It would be nice to make the algorithm take
a TileMap instance in the onTick() method... but that may not work
for all algorithms... ? Only time will reveal whether or not that can
be implemented.

I also made the TILE_SIZE constant, so that it is easier to set the
size of the TileView's in the visualization. The interesting thing is
that when the TILE_SIZE is lower (say 4 or so), it feels as if the
One algorithm is faster, especially at a lower rows/column combination.