Adds the ability for testing in the 'zero algorithm'.
This tesability in zero allows for messing with changes to the visualization
without having to necessarily update the other algorithms at the same time.

This change also makes use of the dirt color, although it looks a little like
something else, but at least it is visually being used, instead of always
having a black background -- which may be useful in the future, to help with
having differently sized fields and viewports.